removed WorldHandler header + minor changes to the class in PerfGrapherCmdlet
